Implemented basic sharding
